Create LICENSE.md
Subject to change to a slightly more restrictive open source license. The objective would be to allow free distribution, forking, modifications, etc. as long as 1. the resulting modified code remains open source and 2. any third party planning to use this project's code base asks for (and receives) permission from the owners and contributors to do so. Please note that the intended conditions in this comment ARE NOT final and do not legally bind any owner, contributor, or user yet.